API

QuranTree.Features — Type

Features(data::String)

Convert a string to morphological feature object.

QuranTree.Lemma — Type

Lemma(data::String)

Convert a string to a Lemma object.

QuranTree.Prefix — Type

Prefix(data::Symbol, pos::AbstractPartOfSpeech)

Create a new Prefix object with data as the symbol of the morphological feature with pos as the corresponding Part of Speech.

QuranTree.Root — Type

Root(data::String)

Convert a string to a Root object.

QuranTree.Special — Type

Special(data::String)

Convert a string to a Special object.

QuranTree.Stem — Type

Stem(data::Symbol, pos::AbstractPartOfSpeech)

Create a new Stem object with data as the symbol of the morphological feature with pos as the corresponding Part of Speech.

QuranTree.Suffix — Type

Suffix(data::Symbol, pos::AbstractPartOfSpeech)

Create a new Suffix object with data as the symbol of the morphological feature with pos as the corresponding Part of Speech.

Base.parse — Method

parse(::Type{Features}, f::AbstractString)

Extract the features of a morphological Feature object.

IndexedTables.table — Method

table(crps::CorpusRaw)

Convert the CorpusRaw read-only array into a tabularized CorpusData using IndexedTable.

IndexedTables.table — Method

table(tnzl::TanzilRaw)

Convert the TanzilRaw read-only array into a tabularized TanzilData using IndexedTable.

QuranTree.arabic — Method

arabic(s::String[, encoder::AbstractEncoder])

Encode the String object into Arabic characters. Custom encoder generated from @transliterator can be provided, but default is Buckwalter.

QuranTree.chapter_name — Function

chapter_name(quran::AbstractQuran, transliterate::Bool=false; lang::Symbol=:arabic)

Extracts the chapter label of the input quran, in either :arabic (default) or :english

QuranTree.dediac — Method

dediac(s::String)

Dediacritize the input String object.

QuranTree.encode — Method

encode(s::String)

Transliterate the input String object using Buckwalter.

QuranTree.encode — Method

encode(::Type{SimpleEncoder}, s::String)

Encode the input String object as SimpleEncoder.

QuranTree.encode — Method

encode(s::Union{Char,String}, encoder::AbstractEncoder)

Transliterate the input String object using a custom encoder. Custom encoder is generated using the @transliterator.

QuranTree.isfeature — Method

isfeature(feat::Features, pos::Type{<:AbstractFeature})

Check if the morphological Feature object is a type of pos.

QuranTree.load — Method

load(data::QuranData)

Load the raw QuranData as a ReadOnlyArray for both Quranic Corpus and Tanzil Data.

QuranTree.normalize — Method

normalize(s::String, chars::Array{Symbol,1})

Normalize a specific Arabic or Buckwalter String character/s (chars).

QuranTree.normalize — Method

normalize(s::String, char::Symbol)

Normalize a specific Arabic or Buckwalter String character (chars).

QuranTree.normalize — Method

normalize(s::String)

Normalize a Arabic or Buckwalter String characters.

QuranTree.verses — Method

verses(data::IndexedTable[, a::Int64[, b::Int64]])

Extract the verses of a IndexedTable object from row a to row b.

QuranTree.verses — Method

verses(quran::AbstractQuran)

Extract the verses of a AbstractQuran object.

QuranTree.verses — Method

verses(quran::TanzilData)

Extract the verses of a TanzilData object.

QuranTree.@data — Macro

@data(expr)

Extract the data property object.

QuranTree.@desc — Macro

@desc(expr)

Extract the detailed description of a AbstractFeature.

QuranTree.@transliterator — Macro

@transliterator(dict, name)

Create a custom transliterator using an input dict (Dict object) with its corresponding name as String object. This will automatically update the transliterator inside all functions like arabic, verses, and encode.

QuranTree.@transliterator — Macro

@transliterator(dict)

Fallback to the default Buckwalter transliterator.